ops/pyproject.toml
[project]
name = "kb-repo-tools"
version = "0.1.0"
description = "Tools for managing the ~/kb knowledge base repository."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "click>=8.1.7",
  "ruamel.yaml>=0.18.6",
  "ulid-py>=1.1.0",
]

[project.scripts]
kb = "kb_repo_tools.cli:main"

[dependency-groups]
dev = [
  "pytest>=8.0.0",
]

[tool.uv]
package = true
